backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
PyJWT
passlib[bcrypt]
python-multipart
psycopg2-binary
pymysql
beautifulsoup4
aiohttp
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
feedparser
duckduckgo-search
apscheduler
alembic
markdown
python-slugify
celery
redis
chromadb
sentence-transformers
playwright
google-api-python-client
google-auth
